+++ api/tests/organizational-entities/acceptance/application/organization/organization.admin.route.test.js
@@ -28,6 +28,71 @@ describe('Acceptance | Organizational Entities | Application | Route | Admin | O
     server = await createServer();
   });
 
+  describe('POST /api/admin/organizations/import-csv', function () {
+    it('create organizations for the given csv file', async function () {
+      // given
+      const superAdminUserId = databaseBuilder.factory.buildUser.withRole().id;
+      databaseBuilder.factory.buildTag({ name: 'GRAS' });
+      databaseBuilder.factory.buildTag({ name: 'GARGOUILLE' });
+      databaseBuilder.factory.buildTag({ name: 'GARBURE' });
+      databaseBuilder.factory.buildFeature(ORGANIZATION_FEATURE.COMPUTE_ORGANIZATION_LEARNER_CERTIFICABILITY);
+      const organizationId = databaseBuilder.factory.buildOrganization().id;
+      const targetProfileId = databaseBuilder.factory.buildTargetProfile({ ownerOrganizationId: organizationId }).id;
+      await databaseBuilder.commit();
+
+      const buffer =
+        'type,externalId,name,provinceCode,credit,createdBy,documentationUrl,identityProviderForCampaigns,isManagingStudents,emailForSCOActivation,DPOFirstName,DPOLastName,DPOEmail,emailInvitations,organizationInvitationRole,locale,tags,targetProfiles\n' +
+        `SCO,ANNEGRAELLE,Orga des Anne-Graelle,33700,666,${superAdminUserId},url.com,,true,,Anne,Graelle,anne-graelle@example.net,,ADMIN,fr,GRAS_GARGOUILLE,${targetProfileId}\n` +
+        `PRO,ANNEGARBURE,Orga des Anne-Garbure,33700,999,${superAdminUserId},,,,,Anne,Garbure,anne-garbure@example.net,,ADMIN,fr,GARBURE,${targetProfileId}`;
+
+      // when
+      const response = await server.inject({
+        method: 'POST',
+        url: `/api/admin/organizations/import-csv`,
+        headers: generateAuthenticatedUserRequestHeaders({ userId: superAdminUserId }),
+        payload: buffer,
+      });
+
+      // then
+      expect(response.statusCode).to.equal(204);
+
+      const organizations = await knex('organizations');
+      expect(organizations).to.have.lengthOf(3);
+
+      const firstOrganizationCreated = organizations.find((organization) => organization.externalId === 'ANNEGRAELLE');
+      expect(firstOrganizationCreated).to.deep.include({
+        type: 'SCO',
+        externalId: 'ANNEGRAELLE',
+        name: 'Orga des Anne-Graelle',
+        provinceCode: '33700',
+        credit: 666,
+        createdBy: superAdminUserId,
+        documentationUrl: 'url.com',
+        identityProviderForCampaigns: null,
+        isManagingStudents: true,
+      });
+
+      const dataProtectionOfficers = await knex('data-protection-officers');
+      expect(dataProtectionOfficers).to.have.lengthOf(2);
+
+      const targetProfileShares = await knex('target-profile-shares');
+      expect(targetProfileShares).to.have.lengthOf(2);
+
+      const firstTargetProfileShare = targetProfileShares.find(
+        (targetProfileShare) => targetProfileShare.organizationId === firstOrganizationCreated.id,
+      );
+      expect(firstTargetProfileShare).to.deep.include({
+        organizationId: firstOrganizationCreated.id,
+        targetProfileId,
+      });
+
+      const firstOrganizationTags = await knex('organization-tags').where({
+        organizationId: firstOrganizationCreated.id,
+      });
+      expect(firstOrganizationTags).to.have.lengthOf(2);
+    });
+  });
+
   describe('GET /api/admin/organizations/{organizationId}/children', function () {
     context('error cases', function () {
       context('when organization does not exist', function () {
